Add files via upload
This is a python program that contains unit tests for the NewTrainer class. The NewTrainer class is a regression training algorithm that is used to train and predict using a dataset. The test_train() function tests the train() method of the NewTrainer class by passing in a sample dataset and asserting that the output is what is expected.